Arreglado el problema de crear archivo con '>>' al inicio, tambien el tema del ctrl+c esta arreglado, ahce algo raro pero funciona, el ctrl+d sigue haciendo el ^C. Ademas he encontrado un buc con las redirecciones no dividiendose bien, ahora ya se dviden bien y se ejecutan correctamente
