+++ Makefile
@@ -6,7 +6,7 @@
 #    By: eros-gir <eros-gir@student.42barcel>       +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2022/08/02 14:48:15 by eros-gir          #+#    #+#              #
-#    Updated: 2023/07/17 20:17:34 by eros-gir         ###   ########.fr        #
+#    Updated: 2023/07/22 16:06:58 by eros-gir         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -27,6 +27,7 @@ SRC = ./src/ms_main.c  \
 	./src/redirections/ms_redirections2.c \
 	./src/redirections/ms_redirections3.c \
 	./src/redirections/ms_heredoc.c \
+	./src/redirections/ms_heredochelpers.c \
 	./src/redirections/ms_redirhelpers.c \
 	./src/parser/env_parser/ms_env_parser.c \
 	./src/tokenizer/ms_tokenizer.c \
