+++ src/executions/ms_pipe.c
@@ -6,7 +6,7 @@
 /*   By: eros-gir <eros-gir@student.42barcel>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/30 18:18:48 by eros-gir          #+#    #+#             */
-/*   Updated: 2023/07/17 20:40:30 by eros-gir         ###   ########.fr       */
+/*   Updated: 2023/07/22 16:23:46 by eros-gir         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -35,6 +35,8 @@ void	msh_pipe_child2(int pobj[2])
 void	msh_pipe_execute(t_vars *vars, t_cmd *tcmd2, t_cmd *tcmd)
 {
 	g_return_status = msh_set_redirect_first(vars, tcmd);
+	if (msh_is_redirect_first(*tcmd))
+		exit (g_return_status);
 	while (msh_is_redirect(*tcmd2))
 	{
 		if (msh_exec_redirect(tcmd2, -1, tcmd2->next->next->argv[0], 0))
