+++ src/redirections/ms_heredoc.c
@@ -6,33 +6,11 @@
 /*   By: eros-gir <eros-gir@student.42barcel>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/30 17:41:42 by eros-gir          #+#    #+#             */
-/*   Updated: 2023/07/21 20:48:46 by eros-gir         ###   ########.fr       */
+/*   Updated: 2023/07/22 16:20:03 by eros-gir         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../incl/mslib.h"
-
-void	msh_close_pipes(int pobj[2])
-{
-	close(pobj[0]);
-	close(pobj[1]);
-}
-
-void	msh_clean_heredoc(t_vars *vars)
-{
-	char	*fname;
-	int		i;
-
-	i = 0;
-	while (i >= 999)
-	{
-		fname = ft_joinloc(ft_strdup(".heredoc"), ft_itoa(i));
-		unlink(fname);
-		free(fname);
-		i++;
-	}
-	vars->hdnumb = 0;
-}
 
 int	msh_store_heredocs(t_vars *vars)
 {
@@ -64,10 +42,10 @@ int	msh_check_sigint(int signum)
 
 	if (signum == 1)
 	{
-		rl_on_new_line();
 		rl_redisplay();
-		write(1, "  \n", 3);
+		write(1, "  ", 2);
 		interrupt = 1;
+		ioctl(0, TIOCSTI, "\n");
 	}
 	else if (signum == -1)
 		interrupt = 0;
@@ -103,8 +81,7 @@ void	msh_heredoc(char *delim, char *fnum)
 		free(line);
 	}
 	msh_check_sigint(-1);
-	signal(SIGINT, SIG_IGN);
-//	signal(SIGINT, msh_sigint_handler);
+	signal(SIGINT, msh_sigint_handler);
 	free(fname);
 	close(fd);
 }
